Installation script is interactive now

+++ install.sh
@@ -2,30 +2,77 @@
 
 dotfiles_dir=$(dirname $(readlink -f $0))
 
-mkdir -p $HOME/.config
+source $dotfiles_dir/prompt-for-multiselect.sh
+
+components=(".bashrc" "Neovim" "Qtile" "AwesomeWM" "Redshift" "Volume control script")
+
+###############################################################################
+# ASK USER WHAT TO INSTALL
+
+bold=$(tput bold)
+normal=$(tput sgr0)
+
+echo -n "Choose config files to install (${bold}SPACE${normal} to select,"
+echo " ${bold}ENTER${normal} to confirm):"
+
+for i in "${!components[@]}"; do
+	options_string+="${components[$i]};"
+done
+
+prompt_for_multiselect SELECTED "$options_string"
+
+for i in "${!SELECTED[@]}"; do
+	if [ "${SELECTED[$i]}" == "true" ]; then
+		checked+=("${components[$i]}")
+	fi
+done
+
+unset bold normal
+
+###############################################################################
+# INSTALLATION
+
+if [[ -z ${checked} ]]; then
+	echo "Nothing was installed."
+	exit 0
+fi
 
 # .bashrc
-rm -f $HOME/.bashrc
-ln -s $dotfiles_dir/.bashrc $HOME
+if [[ " ${checked[@]} " =~ " ${components[0]} " ]]; then
+	ln -sf $dotfiles_dir/.bashrc $HOME
+fi
 
 # Neovim
-rm -f $HOME/.local/share/nvim/site/autoload/plug.vim
-rm -rf $HOME/.config/nvim
-ln -s $dotfiles_dir/.config/nvim $HOME/.config
+if [[ " ${checked[@]} " =~ " ${components[1]} " ]]; then
+	rm -f $HOME/.local/share/nvim/site/autoload/plug.vim
+	mkdir -p $HOME/.config
+	rm -rf $HOME/.config/nvim
+	ln -s $dotfiles_dir/.config/nvim $HOME/.config
+fi
 
 # Qtile
-rm -rf $HOME/.config/qtile
-ln -s $dotfiles_dir/.config/qtile $HOME/.config
+if [[ " ${checked[@]} " =~ " ${components[2]} " ]]; then
+	mkdir -p $HOME/.config
+	rm -rf $HOME/.config/qtile
+	ln -s $dotfiles_dir/.config/qtile $HOME/.config
+fi
 
 # Awesome
-rm -rf $HOME/.config/awesome
-ln -s $dotfiles_dir/.config/awesome $HOME/.config
+if [[ " ${checked[@]} " =~ " ${components[3]} " ]]; then
+	mkdir -p $HOME/.config
+	rm -rf $HOME/.config/awesome
+	ln -s $dotfiles_dir/.config/awesome $HOME/.config
+fi
 
 # Redshift
-rm -rf $HOME/.config/redshift
-ln -s $dotfiles_dir/.config/redshift $HOME/.config
+if [[ " ${checked[@]} " =~ " ${components[4]} " ]]; then
+	mkdir -p $HOME/.config
+	rm -rf $HOME/.config/redshift
+	ln -s $dotfiles_dir/.config/redshift $HOME/.config
+fi
 
 # Volume control script
-mkdir -p $HOME/.local/bin
-rm -rf $HOME/.local/bin/volctl
-ln -s $dotfiles_dir/.local/bin/volctl $HOME/.local/bin
+if [[ " ${checked[@]} " =~ " ${components[5]} " ]]; then
+	mkdir -p $HOME/.local/bin
+	ln -sf $dotfiles_dir/.local/bin/volctl $HOME/.local/bin
+fi
